+++ src/hotspot/cpu/x86/vtableStubs_x86_64.cpp
@@ -118,9 +118,9 @@ VtableStub* VtableStubs::create_vtable_stub(int vtable_index, bool caller_is_c1)
 #ifndef PRODUCT
   if (DebugVtables) {
     Label L;
-    __ cmpptr(method, (int32_t)NULL_WORD);
+    __ cmpptr(method, NULL_WORD);
     __ jcc(Assembler::equal, L);
-    __ cmpptr(Address(method, entry_offset), (int32_t)NULL_WORD);
+    __ cmpptr(Address(method, entry_offset), NULL_WORD);
     __ jcc(Assembler::notZero, L);
     __ stop("Vtable entry is NULL");
     __ bind(L);
@@ -237,9 +237,9 @@ VtableStub* VtableStubs::create_itable_stub(int itable_index, bool caller_is_c1)
 #ifdef ASSERT
   if (DebugVtables) {
     Label L2;
-    __ cmpptr(method, (int32_t)NULL_WORD);
+    __ cmpptr(method, NULL_WORD);
     __ jcc(Assembler::equal, L2);
-    __ cmpptr(Address(method, entry_offset), (int32_t)NULL_WORD);
+    __ cmpptr(Address(method, entry_offset), NULL_WORD);
     __ jcc(Assembler::notZero, L2);
     __ stop("compiler entrypoint is null");
     __ bind(L2);
